Move teleport modal to modals folder, add padding
